refactor: rename FrontPageEnhancer to GalleriesEnhancer
