+++ src/components/Gallery/GalleryEnhancer.vue
@@ -3,8 +3,8 @@ import { computed, onMounted, ref } from 'vue'
 import { VueFinalModal } from 'vue-final-modal'
 import { unsafeWindow } from 'vite-plugin-monkey/dist/client'
 
-import { useGalleryElements } from '@/composables/GalleryEnhancer/useGalleryElements'
-import { usePositions } from '@/composables/GalleryEnhancer/usePositions'
+import { useGalleryElements } from '@/composables/Gallery/useGalleryElements'
+import { usePositions } from '@/composables/Gallery/usePositions'
 import { useWheelStep } from '@/composables/useWheelStep'
 import { useArchive } from '@/composables/useArchive'
 import { useFetchPopups } from '@/composables/useFetchPopups'
@@ -19,7 +19,7 @@ import { fetchAllImages } from '@/utils/fetchImages'
 import { ArchiveDownloadMethod } from '@/constants/monkey'
 import { useFavorite } from '@/composables/useFavorite'
 import { useTorrent } from '@/composables/useTorrent'
-import { useHighlight } from '@/composables/FrontPageEnhancer/useHighlight'
+import { useHighlight } from '@/composables/Galleries/useHighlight'
 import { getElements } from '@/utils/commons'
 
 if (loadAllGalleryImagesSwitch.value) {
