+++ src/components/MultiPageViewer/PageElevator.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { usePages } from '@/composables/MultiPageViewerEnhancer/usePages'
+import { usePages } from '@/composables/MultiPageViewer/usePages'
 
 const {
   pageCount,
