+++ src/components/SettingsPanel/SettingsPanel.vue
@@ -78,7 +78,7 @@ function reload() {
             Scroll by Row
           </h3>
           <span class="settings__notice">
-            *sync with "Front Page Enhancer - Scroll by Row"
+            *sync with "Galleries Enhancer - Scroll by Row"
           </span>
         </div>
 
@@ -287,7 +287,7 @@ function reload() {
 
       <section class="settings-panel__section">
         <h2 class="settings-panel__section-name">
-          Front Page Enhancer
+          Galleries Enhancer
         </h2>
 
         <hr class="settings-panel__separator">
@@ -322,7 +322,7 @@ function reload() {
         <div class="settings">
           <ToggleSwitch v-model="archiveButtonSwitch.value" />
           <h3 class="settings__name">
-            Insert archiver buttons to galleries on the front page.
+            Insert archiver buttons to galleries.
           </h3>
         </div>
       </section>
