+++ src/composables/Galleries/useHighlight.ts
