+++ src/composables/Gallery/useGalleryElements.ts
