+++ src/composables/MultiPageViewer/ImageMagnifier/useMagnifierEvents.ts
