+++ src/composables/MultiPageViewer/ImageMagnifier/useMagnifierGesture.ts
