+++ src/composables/MultiPageViewer/ImageMagnifier/useMagnifierImageLoader.ts
