+++ src/composables/MultiPageViewer/ImageMagnifier/useMagnifierStyle.ts
